# Remove the need for most locking in memory.c.