@oon3m0oo 2018-06-14
I could use some help here since I'm not sure if very old compilers will support __thread and __builtin_expect, but this works great for relatively recent gcc and clang (and should for windows, but I don't have a windows machine handy).