@martin-frbg 2018-06-14
I would draw the line at gcc 4.1 (which came with Red Hat Enterprise Linux 5), which got special treatment for #875. Unfortunately it seems that the Apple version of clang does/did not handle thread_local until XCode 8 or 9. (Which seems to be why two of the travis tests are failing. The third is
a USE_OPENMP=1 build, I am running it locally now to see if it is due to a misplaced ifdef or similar.)