@oon3m0oo 2018-06-14
Given that Apple requires using the latest XCode to release apps to their stores, are we able to assume people will use something recent?

I'll also try an OpenMP build and update the PR if that turns up anything.