@martin-frbg 2018-06-14
USE_OPENMP builds fine, but the automated blas tests crash with
`
OpenBLAS : Program will terminate because you tried to allocate too many memory regions.
BLAS : Bad memory unallocation! :    2  (nil)
`