@oon3m0oo 2018-06-14
My local OpenMP build was fine, and the raw log for the failed build doesn't show any errors, it looks like it's still running?
https://api.travis-ci.org/v3/job/392206901/log.txt