@oon3m0oo 2018-06-14
Ahh... I see.  I suppose setting 2 buffers per thread is probably too small for the tests.  Usually it's num_cpus * 2 * num_threads, but if we have a pool per thread it should be pretty small.  Do you think 4 would be enough?