@oon3m0oo 2018-06-14
Ok I can reproduce this locally with my own program.  I'll dig a bit deeper.