@oon3m0oo 2018-06-14
It appears that when using OpenMP there is only one allocation thread.  I think there's a simple fix.