@oon3m0oo 2018-06-14
Back on those XCode failures it seems it's not an XCode issue, but a target issue.  __thread got support from OSX 10.7, apparently, while OpenBLAS currently targets 10.6, which is now nearly 9 years old.  Is it untenable to increase to 10.7?

Upgrading to a more recent XCode would be nice, but not required.