@martin-frbg 2018-06-14
That's the half-open question from #1580. I*ll do a PR to update the deployment target (10.7 or 10.8 ?)