@oon3m0oo 2018-06-14
I'd certainly be ok with 10.8.  I'm ashamed and proud to admit I have a 2012 Macbook Pro laptop at home and it's running just fine on 10.13.  Apple has a pretty good track record of supporting old non-phone hardware:
https://support.apple.com/kb/SP765?locale=en_US